+++ lib/src/countly_flutter.dart
@@ -2209,6 +2209,11 @@ class Countly {
         countlyConfig['requestDropAgeHours'] = config.requestDropAgeHours;
       }
 
+      if (config.providedServerConfig != null) {
+        log('"_configToJson", value provided for providedServerConfig: [${config.providedServerConfig}]', logLevel: LogLevel.INFO);
+        countlyConfig['providedServerConfig'] = config.providedServerConfig;
+      }
+
       /// Experimental ---------------------------
       if (config.experimental.visibilityTracking) {
         log('"_configToJson", value provided for visibilityTracking: [${config.experimental.visibilityTracking}]', logLevel: LogLevel.INFO);